test: move once connection tests from aw_test.go to once_test.go and update them to use OncePool
